% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybootpval.R
\name{bootpval}
\alias{bootpval}
\title{Bootstrap function with graph display}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{sample}

\item{conf.level}{confidence level}

\item{iter}{total number of iterations}

\item{mu0}{mean if the first sample if two are present}

\item{test}{the kind of test being done, either one or two with a default of two}
}
\value{
histogram of accepted and rejected regions for a given mean and sample.
}
\description{
Function from lab 12 that displayus a histogram and accepted and rejected regions for a given sample of data
}
\examples{
\dontrun{set.seed(55);x1=rnorm(30,mean=25, sd=5),bootpval(x=x1,mu=25,test="two")}

}
