% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{myboot2}
\alias{myboot2}
\title{bootstrap function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05)
}
\arguments{
\item{iter}{number of iterations performed}

\item{x}{sample size}

\item{fun}{Type of calculation performed}

\item{alpha}{Confidence interval wanted}
}
\value{
Plot of the bootstrapped sample with labled bounds of the specified confidence interval
}
\description{
Function from lab 9 that bootstraps samples and calculates confidence intervals
}
\examples{
\dontrun{set.seed(68); sam=rnorm(20,mean=10,sd=4); myboot2(x=sam,col="yellow")}
}
