% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myncurve simulation}
\usage{
myci(x)
}
\arguments{
\item{x}{random sample}
}
\value{
left and right bounds of the 95% confidence interval about the mean
}
\description{
Function from lab 11 that calculates the 95% confidence interval
}
\examples{
\dontrun{myci(x=rnorm(30,mean=10,sd=3))}
}
