% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_clt.R
\name{mycltp}
\alias{mycltp}
\title{Multidimenstional simulation}
\usage{
mycltp(n = 10, iter = 10000, lambda = 10)
}
\arguments{
\item{n}{number of trials}

\item{iter}{number of iterations}

\item{lambda}{Parameter}
}
\value{
3 panel graphs for the poisson distribution
}
\description{
Poisson Central Limit Thereom function from lab 8
}
\examples{
\dontrun{mycltp(n = 2,iter = 10000,lambda=4)}
}
