% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Multidimenstional simulation}
\usage{
mymlnorm(
  x,
  mu = seq(10, 15, length = 1000),
  sig = seq(0.1, 4, length = 1000),
  color = "green"
)
}
\arguments{
\item{x}{vector of data}

\item{mu}{mean of the sample}

\item{sig}{standard deviation of the sample}

\item{color}{Color of plot}

\item{length}{iterations}
}
\value{
3 panel graphs for the poisson distribution
}
\description{
Poisson Central Limit Thereom function from lab 8
}
\examples{
\dontrun{y=c(10,12,13,15,12,11,10); mymlnorm(y)}

}
