% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.R
\name{mymult}
\alias{mymult}
\title{Multidimenstional simulation}
\usage{
mymult(iter = 100, n = 10, p = c(1, 1, 1, 1)/4)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{number of trials}

\item{p}{probability vector}
}
\value{
barplot and table of relative frequencies
}
\description{
Function from lab 5
}
\examples{
\dontrun{mymult()}
}
