% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve simulation}
\usage{
myncurve(a = 10, mu = 3, sigma = 1)
}
\arguments{
\item{a}{point on the curve}

\item{mu}{mean of the distribution}

\item{sigma}{standard deviation}
}
\value{
curve of shaded polygon and probabiities
}
\description{
Function from lab 6 that calculates probability from -inf to point a
}
\examples{
\dontrun{myncurve(a=10,mu=3,sigma=1)}
}
