% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package1.R
\name{myquadratic}
\alias{myquadratic}
\title{Lab 4 function}
\usage{
myquadratic(x = 10)
}
\arguments{
\item{x}{is an independent value}
}
\value{
yhat value
}
\description{
Lab 4 function
}
\examples{
\dontrun{myplot(x=13)}

}
